Linked List - 1
Polynomial Using LinkedList
Polynomial are abstract mathematical concept generally used in algebra. A polynomial comprises of terms and each term is defined as coefficent, variable and exponents of the variable. For example: 3X^2 is a term in which 3 is a coefficieent, X is a variable and 2 is the exponent. A polynomial can be comprises of many terms. You will be entered the number of terms in first polynomial and after reading it, you will be reading each term of the polynomial as a triplet, for example 3 X 2. After reading first polynimal, you can read the second polynomial in the similar fashion. These polynomial terms are not in sorted order of their exponents. You are required to create a linked list of your choice to process these two polynomails. After reading these polynomials you are required to perform addition of these polynomials, after that you are required to write the resultant polynomial in increasing order of exponents for each variable.